Feat unifier layout et navigation
