+++ app/ScrollSectionsWrapper.jsx
@@ -1,5 +1,5 @@
 "use client";
-import { useScrollAnchors } from "../src/hooks/useScrollAnchors";
+import { useScrollAnchors } from "../src/hooks";
 import { sections } from "../src/assets/data/sections";
 
 const ScrollSectionsWrapper = ({ children }) => {
