+++ src/components/header/HeaderLazy.jsx
@@ -1,14 +1,13 @@
 "use client";
 import dynamic from "next/dynamic";
 // import HeaderGhost from "./HeaderGhost";
-
-const HeaderWarpProvider = dynamic(() => import("./HeaderWarpProvider"), {
+const Header = dynamic(() => import("./Header"), {
     ssr: false,
     // loading: () => <HeaderGhost />,
 });
 
 const HeaderLazy = () => {
-    return <HeaderWarpProvider />;
+    return <Header />;
 };
 
 export default HeaderLazy;
