--- src/components/header/HeaderWarpProvider.jsx
@@ -1,17 +1,0 @@
-"use client";
-import React from "react";
-import { NavigationProvider } from "../../utils/context/NavigationContext";
-import dynamic from "next/dynamic";
-const Header = dynamic(() => import("./Header"), {
-    ssr: false,
-});
-
-const HeaderWarpProvider = () => {
-    return (
-        <NavigationProvider>
-            <Header />
-        </NavigationProvider>
-    );
-};
-
-export default React.memo(HeaderWarpProvider);
