+++ src/workers/index.ts
@@ -1,0 +1,2 @@
+export const SCROLL_WORKER = "/public/workers/scrollWorker.js";
+export const SCROLL_SMOOTH_WORKER = "/public/workers/scrollSmoothWorker.js";
